Move NestedSortBuilder to its own class.
