+++ core/src/main/java/org/elasticsearch/search/sort/SortBuilder.java
@@ -29,12 +29,8 @@ import org.elasticsearch.common.ParseField;
 import org.elasticsearch.common.ParsingException;
 import org.elasticsearch.common.Strings;
 import org.elasticsearch.common.io.stream.NamedWriteable;
-import org.elasticsearch.common.io.stream.StreamInput;
-import org.elasticsearch.common.io.stream.StreamOutput;
-import org.elasticsearch.common.io.stream.Writeable;
 import org.elasticsearch.common.lucene.search.Queries;
 import org.elasticsearch.common.xcontent.ToXContentObject;
-import org.elasticsearch.common.xcontent.XContentBuilder;
 import org.elasticsearch.common.xcontent.XContentParser;
 import org.elasticsearch.index.fielddata.IndexFieldData.XFieldComparatorSource.Nested;
 import org.elasticsearch.index.mapper.ObjectMapper;
@@ -63,9 +59,6 @@ public abstract class SortBuilder<T extends SortBuilder<T>> implements NamedWrit
     public static final ParseField ORDER_FIELD = new ParseField("order");
     public static final ParseField NESTED_FILTER_FIELD = new ParseField("nested_filter");
     public static final ParseField NESTED_PATH_FIELD = new ParseField("nested_path");
-    public static final ParseField NESTED_FIELD = new ParseField("nested");
-    public static final ParseField PATH_FIELD = new ParseField("path");
-    public static final ParseField FILTER_FIELD = new ParseField("filter");
 
     private static final Map<String, Parser<?>> PARSERS;
     static {
@@ -76,132 +69,6 @@ public abstract class SortBuilder<T extends SortBuilder<T>> implements NamedWrit
         parsers.put(ScoreSortBuilder.NAME, ScoreSortBuilder::fromXContent);
         // FieldSortBuilder gets involved if the user specifies a name that isn't one of these.
         PARSERS = unmodifiableMap(parsers);
-    }
-
-    public static class NestedSortBuilder implements Writeable, Reader<NestedSortBuilder>, ToXContentObject {
-        private String path;
-        private QueryBuilder filter;
-        private NestedSortBuilder nestedSort;
-
-        public NestedSortBuilder() {
-        }
-
-        public NestedSortBuilder(StreamInput in) throws IOException {
-            read(in);
-        }
-
-        public String getPath() {
-            return path;
-        }
-
-        public NestedSortBuilder setPath(final String path) {
-            this.path = path;
-            return this;
-        }
-
-        public QueryBuilder getFilter() {
-            return filter;
-        }
-
-        public NestedSortBuilder setFilter(final QueryBuilder filter) {
-            this.filter = filter;
-            return this;
-        }
-
-        public NestedSortBuilder getNestedSort() {
-            return nestedSort;
-        }
-
-        public NestedSortBuilder setNestedSort(final NestedSortBuilder nestedSortBuilder) {
-            this.nestedSort = nestedSortBuilder;
-            return this;
-        }
-
-        /**
-         * Write this object's fields to a {@linkplain StreamOutput}.
-         */
-        @Override
-        public void writeTo(final StreamOutput out) throws IOException {
-            out.writeOptionalString(path);
-            out.writeOptionalNamedWriteable(filter);
-            out.writeOptionalWriteable(nestedSort);
-        }
-
-        /**
-         * Read {@code V}-type value from a stream.
-         *
-         * @param in Input to read the value from
-         */
-        @Override
-        public NestedSortBuilder read(final StreamInput in) throws IOException {
-            path = in.readOptionalString();
-            filter = in.readOptionalNamedWriteable(QueryBuilder.class);
-            nestedSort = in.readOptionalWriteable(NestedSortBuilder::new);
-            return this;
-        }
-
-        @Override
-        public XContentBuilder toXContent(final XContentBuilder builder, final Params params) throws IOException {
-            builder.startObject();
-            if (path != null) {
-                builder.field(PATH_FIELD.getPreferredName(), path);
-            }
-            if (filter != null) {
-                builder.field(FILTER_FIELD.getPreferredName(), filter);
-            }
-            if (nestedSort != null) {
-                builder.field(NESTED_FIELD.getPreferredName(), nestedSort);
-            }
-            builder.endObject();
-            return builder;
-        }
-
-        public static NestedSortBuilder fromXContent(XContentParser parser) throws IOException {
-            NestedSortBuilder nestedSortBuilder = new NestedSortBuilder();
-            XContentParser.Token token = parser.currentToken();
-            if (token == XContentParser.Token.START_OBJECT) {
-                while ((token = parser.nextToken()) != XContentParser.Token.END_OBJECT) {
-                    if (token == XContentParser.Token.FIELD_NAME) {
-                        String currentName = parser.currentName();
-                        parser.nextToken();
-                        if (currentName.equals(PATH_FIELD.getPreferredName())) {
-                            nestedSortBuilder.setPath(parser.text());
-                        } else if (currentName.equals(FILTER_FIELD.getPreferredName())) {
-                            nestedSortBuilder.setFilter(parseNestedFilter(parser));
-                        } else if (currentName.equals(NESTED_FIELD.getPreferredName())) {
-                            nestedSortBuilder.setNestedSort(NestedSortBuilder.fromXContent(parser));
-                        } else {
-                            throw new IllegalArgumentException("malformed nested sort format, unknown field name [" + currentName + "]");
-                        }
-                    } else {
-                        throw new IllegalArgumentException("malformed nested sort format, only field names are allowed");
-                    }
-                }
-            } else {
-                throw new IllegalArgumentException("malformed nested sort format, must start with an object");
-            }
-
-            return nestedSortBuilder;
-        }
-
-        @Override
-        public boolean equals(final Object obj) {
-            if (this == obj) {
-                return true;
-            }
-            if (obj == null || getClass() != obj.getClass()) {
-                return false;
-            }
-            NestedSortBuilder that = (NestedSortBuilder) obj;
-            return Objects.equals(path, that.path)
-                && Objects.equals(filter, that.filter)
-                && Objects.equals(nestedSort, that.nestedSort);
-        }
-
-        @Override
-        public int hashCode() {
-            return Objects.hash(path, filter, nestedSort);
-        }
     }
 
     /**
